fix unicode handling of response from eLife XML API
The response coming from requests library is a string of bytes - there is no
need to do .encode() on it.

encode() converts a Python Unicode thingy into a string of bytes. We need to
DEcode the string of bytes coming in from requests library.

DEcoding turns a string of bytes into a Python Unicode thingy. If you try to
encode() a string of bytes, it tries to automatically DEcode() it first using
the default encoding, ascii. If the response has utf-8 characters in it, you
get an error telling you "ascii codec does not know how to decode character at
<position>".