Change the way plugins declare their name (+tests)
The PLoS plugin had the _short_name of 'oup' which meant that all PLoS
items got identified as Oxford University Press! The reason for that was
that the plugin name was stated explicitly, instead of being the same as
the module (plugin file) name.

Also, the tests were kind of pointless - they tested whether the
_short_name declared in the plugin was what got recorded in the
'handler' provenance field. The tests should test explicitly for the
correct plugin name and version.

Both problems have been fixed across all plugins.